[FIX] hr_attendance: department employee count in kiosk
**PROBLEM**
When using the kiosk on a company that's not the default company, the employee count displayed for all department is 0.

**STEP TO REPRODUCE**
1.install the attendance module.
2.select another company.
3.go to the kiosk (attendance/Kiosk Mode).
4.identify manually.
5.notice the department count is incorrectly displayed as 0.

**CAUSE**
When entering the kiosk, the user is disconnected from their session, the user in this context is the public_user who have only access to the default company.
The _compute_total_employee filter employees based on the company we have access to, so any employee from the non-default company are not taken into account.

**FIX**
We set the allowed_company_ids to the correct company id in the request context.

opw-4647574

closes odoo/odoo#213935

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>