[FIX] l10n_jo_edi: Credit notes lines matching by discount
When generating a credit note for an invoice with multiple lines of the same product
(same product_id, name, and price_unit) but different discounts,
the system incorrectly matches both credit note lines to the same invoice line.
This commit solves this issue by adding line discount to
the criteria used to match credit notes lines with invoices lines.

task-4876849

closes odoo/odoo#215402

X-original-commit: d86409d93e096126b32cbb35f442db332628375d
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Hesham Saleh (hsal) <hsal@odoo.com>