[FIX] web: Hoot - allow to exclude jobs from UI
Before this commit, it was not possible to toggle "off" a suite, test
or tag from the search bar due to an incorrect handling of the internal
value given by input handlers.

This has been fixed in this commit, allowing to exclude jobs from UI
again.

Part-of: odoo/odoo#213223
Related: odoo/enterprise#87476
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>