[FIX] test_website: update timeout for testing
## Version
18.0+

## Issue
Task 4141409 introduced long tests to simulate complete flow. As this test relies on many iframe checks and these iframes being slow to load, the test time limit is reached.
The test fails on Single App and Community but not with Enterprise might be related to performances linked to modules interactions.

runbot-223225

closes odoo/odoo#214870

Signed-off-by: Hubert Van De Walle <huvw@odoo.com>